backend/requirements.txt
feedparser 
google-generativeai
python-dotenv